pyproject.toml
[tool.poetry]
name = "resumedataapi"
version = "0.1.0"
description = ""
authors = ["ApprenticeofEnder <perfknight@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.10"
django = "^4.0.5"
djangorestframework = "^3.13.1"